Adicionado transformers

+++ app/Http/Controllers/ProjectController.php
@@ -9,16 +9,19 @@ use Dashboard\Http\Requests;
 use Dashboard\Repositories\ProjectRepository;
 use Dashboard\Services\ProjectService;
 
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
+
 class ProjectController extends Controller
 {
     private $repository;
     private $service;
+    private $userId;
     
     public function __construct(ProjectRepository $repository, ProjectService $service)
     {
         $this->repository = $repository;
-        
         $this->service = $service;
+        $this->userId = Authorizer::getResourceOwnerId();
     }
     
     /**
@@ -28,7 +31,7 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        return $this->repository->with(['owner', 'client'])->all();
+        return $this->repository->with(['owner', 'client'])->findWhere(['owner_id' => $this->userId]);
     }
 
     /**
@@ -50,7 +53,11 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
-        return $this->repository->with(['owner', 'client'])->find($id);
+        if($this->checkPrivileges($id))
+        {
+            return $this->repository->with(['owner', 'client'])->find($id);
+        }
+        return ['error' => 'You do not have access rights for this project'];
     }
 
     /**
@@ -62,7 +69,11 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->service->update($request->all(), $id);
+        if($this->checkPrivileges($id))
+        {
+            $this->service->update($request->all(), $id);
+        }
+        return ['error' => 'You do not have access rights for this project'];
     }
 
     /**
@@ -73,6 +84,30 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
-        $this->repository->delete($id);
+        if($this->checkIsOwner($id))
+        {
+            $this->repository->delete($id);
+        }
+        return ['error' => 'You do not have access rights to finish this action'];
+    }
+    
+    /**
+    *
+    * Authorization functions.
+    */
+    
+    private function checkIsOwner($projectId)
+    {   
+        return $this->repository->isOwner($projectId, $this->userId);
+    }
+    
+    private function checkIsMember($projectId)
+    {
+        return $this->repository->isMember($projectId, $this->userId);
+    }
+    
+    private function checkPrivileges($projectId)
+    {
+        return ($this->checkIsOwner($projectId) || $this->checkIsMember($projectId)) ? true : false;
     }
 }
